requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
blis==1.0.1
Brlapi==0.8.3
catalogue==2.0.10
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
cliapp==1.20180812.1
click==8.0.3
cloudpathlib==0.19.0
cmdtest==0.32+git
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
confection==0.1.5
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.8
dbus-python==1.2.18
debugpy==1.8.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
duplicity==0.8.21
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
future==0.18.2
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.67.1
gyp==0.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.26.0
ipython_genutils==0.2.0
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server-mathjax==0.2.6
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_git==0.50.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keras==3.6.0
keyring==23.5.0
kiwisolver==1.4.5
langcodes==3.4.1
language-selector==0.1
language_data==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
marisa-trie==1.2.0
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
monotonic==1.6
more-itertools==8.10.0
murmurhash==1.0.10
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbdime==4.0.1
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
nltk==3.9.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.2
oauthlib==3.2.0
olefile==0.46
opt_einsum==3.4.0
optree==0.13.0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.2
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.9.11
reportlab==3.6.8
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.19.0
scikit-learn==1.5.2
scipy==1.14.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.2
srsly==2.4.8
stack-data==0.6.3
systemd-python==234
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
textblob==0.18.0.post0
thinc==8.3.2
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
ttystatus==0.38
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.27.0
wadllib==1.3.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.1
wordcloud==1.9.3
wrapt==1.16.0
xdg==5
xkit==0.0.0
zipp==1.0.0
